Update from origin

+++ concrete/src/File/Set/Set.php
@@ -1,17 +1,24 @@
 <?php
 namespace Concrete\Core\File\Set;
 
+use Concrete\Core\Database\Connection\Connection;
+use Concrete\Core\Entity\File\File as FileEntity;
+use Concrete\Core\Entity\File\Version as FileVersionEntity;
+use Concrete\Core\Support\Facade\Application;
 use Concrete\Core\Permission\Access\Entity\GroupCombinationEntity as GroupCombinationPermissionAccessEntity;
 use Concrete\Core\Permission\Access\Entity\GroupEntity as GroupPermissionAccessEntity;
 use Concrete\Core\Permission\Access\Entity\UserEntity as UserPermissionAccessEntity;
 use Concrete\Core\Permission\Key\FileSetKey as FileSetPermissionKey;
+use Concrete\Core\Support\Facade\Facade;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Events;
 use File as ConcreteFile;
 use Database;
 use Concrete\Core\Permission\Access\Access as PermissionAccess;
 use PermissionKey;
 use Permissions;
-use User;
+use Concrete\Core\User\User;
 
 /**
  * Represents a file set.
@@ -69,28 +76,78 @@ class Set
     }
 
     /**
-     * @param bool|\User $u
+     * Returns all sets currently available to the User
+     *
+     * @param bool|User|\Concrete\Core\User\UserInfo $user
+     *
+     * @return static[]
+     */
+    public static function getMySets($user = false)
+    {
+        $app = Facade::getFacadeApplication();
+
+        if ($user === false) {
+            $user = $app->make(User::class);
+        }
+
+        /** @var $database \Concrete\Core\Database\Connection\Connection */
+        $database = $app->make('database')->connection();
+        $fileSets = array();
+
+        $queryBuilder = $database->createQueryBuilder();
+        $results = $queryBuilder->select('*')->from('FileSets')->where(
+            $queryBuilder->expr()->eq('fsType', self::TYPE_PUBLIC)
+            )->orWhere(
+                $queryBuilder->expr()->andX(
+                    $queryBuilder->expr()->in('fsType',[self::TYPE_PRIVATE, self::TYPE_STARRED, self::TYPE_PUBLIC]),
+                    $queryBuilder->expr()->eq('uID', $user->getUserID())
+                )
+            )->execute();
+
+
+        while ($row = $results->fetch()) {
+            $fileSet = new static();
+            $fileSet = array_to_object($fileSet, $row);
+            $fileSets[] = $fileSet;
+        }
+
+        return $fileSets;
+    }
+
+    /**
+     * Returns all sets (except saved searches) for a User
+     *
+     * @param bool|User|\Concrete\Core\User\UserInfo $user User or UserInfo Object
      *
      * @return array
      */
-    public static function getMySets($u = false)
+    public static function getOwnedSets($user = false)
     {
-        if ($u === false) {
-            $u = new User();
-        }
-        $db = Database::connection();
-        $sets = array();
-        $r = $db->executeQuery(
-            'SELECT * FROM FileSets WHERE fsType = ? OR (fsType IN (?, ?) AND uID = ?) ORDER BY fsName ASC',
-            array(self::TYPE_PUBLIC, self::TYPE_STARRED, self::TYPE_PRIVATE, $u->getUserID())
-        );
-        while ($row = $r->fetch()) {
-            $fs = new static();
-            $fs = array_to_object($fs, $row);
-            $sets[] = $fs;
+
+        $app = Facade::getFacadeApplication();
+
+        if ($user === false) {
+            $user = $app->make(User::class);
         }
 
-        return $sets;
+        /** @var \Concrete\Core\Database\Connection\Connection $database */
+        $database = $app->make('database')->connection();
+        $fileSets = array();
+
+        $queryBuilder = $database->createQueryBuilder();
+        $results = $queryBuilder->select('*')->from('FileSets')->where(
+            $queryBuilder->expr()->in('fsType',[self::TYPE_PRIVATE, self::TYPE_STARRED, self::TYPE_PUBLIC])
+        )->andWhere($queryBuilder->expr()->eq('uID', $user->getUserID()))->execute();
+
+
+        while ($row = $results->fetch()) {
+            $fileSet = new static();
+            $fileSet = array_to_object($fileSet, $row);
+            $fileSets[] = $fileSet;
+        }
+
+        return $fileSets;
+
     }
 
     /**
@@ -407,21 +464,43 @@ class Set
     /**
      * Adds the file to the set.
      *
-     * @param int|\File $f_id //accepts an ID or a File object
+     * @param int|FileEntity|FileVersionEntity $f_id
      *
-     * @return File|mixed
+     * @return \Concrete\Core\File\Set\File|null returns NULL if the operation failed (for instance because $f_id is invalid), a \Concrete\Core\File\Set\File instance otherwise  
      */
     public function addFileToSet($f_id)
     {
+        $app = Application::getFacadeApplication();
         if (is_object($f_id)) {
-            $f_id = $f_id->getFileID();
+            $f = $f_id;
+            if ($f instanceof FileEntity) {
+                $file = $f;
+                $fileVersion = $file->getApprovedVersion();
+            } else {
+                $fileVersion = $f;
+                $file = $fileVersion->getFile();
+            }
+            $f_id = (int) $file->getFileID();
+        } else {
+            $f_id = (int) $f_id;
+            $em = $app->make(EntityManagerInterface::class);
+            $file = $em->find(FileEntity::class, $f_id);
+            $fileVersion = $file->getApprovedVersion();
         }
-        $file_set_file = File::createAndGetFile($f_id, $this->fsID);
+        if ($file === null) {
+            $result = null;
+        } else {
+            $file_set_file = File::createAndGetFile($f_id, $this->fsID);
+            $fe = new \Concrete\Core\File\Event\FileSetFile($file_set_file);
+            $director = $app->make(EventDispatcherInterface::class);
+            $director->dispatch('on_file_added_to_set', $fe);
+            if ($fileVersion !== null) {
+                $fileVersion->refreshThumbnails(false);
+            }
+            $result = $file_set_file;
+        }
 
-        $fe = new \Concrete\Core\File\Event\FileSetFile($file_set_file);
-        Events::dispatch('on_file_added_to_set', $fe);
-
-        return $file_set_file;
+        return $result;
     }
 
     public function getSavedSearchRequest()
@@ -435,23 +514,48 @@ class Set
     }
 
     /**
-     * @param int|\File $f_id
+     * @param int|FileEntity|FileVersionEntity $f_id
+     *
+     * @return bool Returns false if the operation failed (for instance because $f_id is invalid), true otherwise
      */
     public function removeFileFromSet($f_id)
     {
+        $app = Application::getFacadeApplication();
         if (is_object($f_id)) {
-            $f_id = $f_id->getFileID();
+            $f = $f_id;
+            if ($f instanceof FileEntity) {
+                $file = $f;
+                $fileVersion = $file->getApprovedVersion();
+            } else {
+                $fileVersion = $f;
+                $file = $fileVersion->getFile();
+            }
+            $f_id = (int) $file->getFileID();
+        } else {
+            $f_id = (int) $f_id;
+            $em = $app->make(EntityManagerInterface::class);
+            $file = $em->find(FileEntity::class, $f_id);
+            $fileVersion = $file->getApprovedVersion();
+        }
+        if ($file === null) {
+            $result = false;
+        } else {
+            $file_set_file = File::createAndGetFile($f_id, $this->fsID);
+            $db = $app->make(Connection::class);
+            $db->executeQuery(
+                'DELETE FROM FileSetFiles WHERE fID = ? AND fsID = ?',
+                [$f_id, $this->getFileSetID()]
+            );
+            $fe = new \Concrete\Core\File\Event\FileSetFile($file_set_file);
+            $director = $app->make(EventDispatcherInterface::class);
+            $director->dispatch('on_file_removed_from_set', $fe);
+            if ($fileVersion !== null) {
+                $fileVersion->refreshThumbnails(false);
+            }
+            $result = true;
         }
 
-        $file_set_file = File::createAndGetFile($f_id, $this->fsID);
-
-        $db = Database::connection();
-        $db->executeQuery(
-            'DELETE FROM FileSetFiles WHERE fID = ? AND fsID = ?',
-            array($f_id, $this->getFileSetID())
-        );
-        $fe = new \Concrete\Core\File\Event\FileSetFile($file_set_file);
-        Events::dispatch('on_file_removed_from_set', $fe);
+        return $result;
     }
 
     public function hasFileID($f_id)
@@ -468,9 +572,14 @@ class Set
 
     public function delete()
     {
+        $fe = new \Concrete\Core\File\Event\FileSet($this);
+        Events::dispatch('on_file_set_delete', $fe);
+
         $db = Database::connection();
         $db->delete('FileSets', array('fsID' => $this->fsID));
         $db->executeQuery('DELETE FROM FileSetSavedSearches WHERE fsID = ?', array($this->fsID));
+        $db->executeQuery('DELETE FROM FileSetFiles WHERE fsID = ?', array($this->fsID));
+        $db->executeQuery('DELETE FROM FileImageThumbnailTypeFileSets WHERE ftfsFileSetID = ?', [$this->fsID]);
     }
 
     /*
